0.1.3
0.1.3

+++ src/main/java/io/tbbc/cf/modifier/Modifier.java
@@ -1,4 +1,4 @@
 package io.tbbc.cf.modifier;
 
-public record Modifier(ModifierName name, long value) {
+public record Modifier(ModifierName name, int value) {
 }
